requirements.txt
fastai==1.0.61
numpy==1.17.2
pandas==1.1.2 
torch==1.6.0
torchvision===0.7.0
psycopg2-binary==2.8.6



